moving FeideApiClient from learningpath-api to network
