+++ learningpath-api/src/main/scala/no/ndla/learningpathapi/service/ReadService.scala
@@ -9,7 +9,6 @@
 package no.ndla.learningpathapi.service
 
 import cats.implicits._
-import no.ndla.learningpathapi.integration.FeideApiClient
 import no.ndla.learningpathapi.model.api
 import no.ndla.learningpathapi.model.api._
 import no.ndla.learningpathapi.model.domain
@@ -23,6 +22,7 @@ import no.ndla.learningpathapi.model.domain.{
   _
 }
 import no.ndla.learningpathapi.repository.{ConfigRepository, FolderRepository, LearningPathRepositoryComponent}
+import no.ndla.network.clients.FeideApiClient
 import scalikejdbc.DBSession
 
 import java.util.UUID
