+++ learningpath-api/src/main/scala/no/ndla/learningpathapi/service/UpdateService.scala
@@ -9,7 +9,7 @@
 package no.ndla.learningpathapi.service
 
 import no.ndla.learningpathapi.Props
-import no.ndla.learningpathapi.integration.{FeideApiClient, SearchApiClient, TaxonomyApiClient}
+import no.ndla.learningpathapi.integration.{SearchApiClient, TaxonomyApiClient}
 import no.ndla.learningpathapi.model.api.config.UpdateConfigValue
 import no.ndla.learningpathapi.model.api.{config, _}
 import no.ndla.learningpathapi.model.api
@@ -21,6 +21,7 @@ import no.ndla.learningpathapi.service.search.SearchIndexService
 import no.ndla.learningpathapi.validation.{LearningPathValidator, LearningStepValidator}
 import cats.implicits._
 import no.ndla.common.Clock
+import no.ndla.network.clients.FeideApiClient
 import scalikejdbc.{DBSession, ReadOnlyAutoSession}
 
 import java.util.UUID
