+++ learningpath-api/src/test/scala/no/ndla/learningpathapi/TestEnvironment.scala
@@ -27,6 +27,7 @@ import no.ndla.learningpathapi.service._
 import no.ndla.learningpathapi.service.search.{SearchConverterServiceComponent, SearchIndexService, SearchService}
 import no.ndla.learningpathapi.validation._
 import no.ndla.network.NdlaClient
+import no.ndla.network.clients.FeideApiClient
 import no.ndla.search.{BaseIndexService, Elastic4sClient, NdlaE4sClient}
 import org.mockito.scalatest.MockitoSugar
