+++ network/src/main/scala/no/ndla/network/model/package.scala
@@ -1,0 +1,13 @@
+/*
+ * Part of NDLA network.
+ * Copyright (C) 2022 NDLA
+ *
+ * See LICENSE
+ */
+
+package no.ndla.network
+
+package object model {
+  type FeideID          = String
+  type FeideAccessToken = String
+}
